FIX PostService 권한 확인 로직 추가

+++ src/main/java/com/spring/clone/post/PostService.java
@@ -5,6 +5,7 @@ import com.spring.clone.post.dto.PostRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -22,18 +23,26 @@ public class PostService {
         postRepository.save(newPost);
     }
 
-    public void editPost(Long postId, PostRequestDto requestDto) {
+    public void editPost(Long postId, PostRequestDto requestDto, User user) {
         Post post = getPostById(postId);
-        post.update(requestDto);
-        postRepository.save(post);
+        if ( post.isWritedBy(user) ) {
+            post.update(requestDto);
+            postRepository.save(post);
+        } else {
+            throw new AccessDeniedException("권한이 없습니다.");
+        }
     }
 
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId, User user) {
         Post post = getPostById(postId);
-        postRepository.delete(post);
+        if ( post.isWritedBy(user) ) {
+            postRepository.delete(post);
+        } else {
+            throw new AccessDeniedException("권한이 없습니다.");
+        }
     }
 
-    private Post getPostById(Long postId) {
+    public Post getPostById(Long postId) {
         return postRepository.findById(postId).orElseThrow(
                 () -> new PostNotFoundException("게시글을 찾을 수 없습니다.")
         );
